refactored a method signature in IDataAccessController interface. Completed DataAccessController class. Added source files for controller tests.

+++ src/gissystem/interfaces/IDataAccessController.java
@@ -5,7 +5,7 @@ import gissystem.controllers.QuadTreeController;
 
 public interface IDataAccessController {
 	public String getNextCommandLine();
-	public String findInDatabase( Integer offset );
+	public String findInDatabase( Long offset );
 	public QuadTreeController getQuadTreeController();
 	public HashTableController getHashTableController();
 	public ILogger getLogger();
